Fixed GPS location not showing
The GPS location was not showing up when the coordinates resulted in being in a different UTM zone compared to the project's zone. For example for a project in zone 55, if longitude was > 150 degrees it would assume zone 56 and return an easting/northing value that placed the marker off the map.